+++ CHANGELOG.md
@@ -2409,3 +2409,4 @@
 - [2026-06-29 12:59] feat(analytics): add bi-directional event bus handler for analytics
 - [2026-06-29 13:00] style(analytics): harmonize light and dark theme palette for analytics
 - [2026-06-29 16:26] feat(test): add async test stream provider
+- [2026-06-29 16:35] test(storage): add unit tests for storage business logic